feat: refactor after code review, DEV-10-button
